Move widget to it's own file for tests
